change test song to compare against mognotate.
results: twice as fast.

+++ BardMusicPlayer.Notate/Song/BmpSong.cs
@@ -37,6 +37,6 @@ namespace BardMusicPlayer.Notate.Song
         /// <summary>
         /// 
         /// </summary>
-        public Dictionary<int, TrackContainer> TrackContainers { get; set; } = new();
+        public Dictionary<long, TrackContainer> TrackContainers { get; set; } = new();
     }
 }
